VIX-2806 Performance optimizations.
Change the logic to reuse the temp pixel buffer instead of recreating it every cycle. From memory perspective updating the values in the existing buffer is faster than allocating a new one. There is a clear method on the buffer than can be utilized for this.

Make a change in the ClearBuffer method to optimize how fast the arrays can be initialized. This uses some tricks with Array.Copy that utilizes underlying system calls that are faster.